+++ packages/app/src/flame/variations/simple/general/flipCircleVar.ts
@@ -1,4 +1,4 @@
-import { vec2f } from 'typegpu/data'
+import { f32, vec2f } from 'typegpu/data'
 import { select } from 'typegpu/std'
 import { simpleVariation } from '../types'
 
@@ -8,7 +8,7 @@ export const flipCircleVar = simpleVariation(
     'use gpu'
     const r2 = pos.x * pos.x + pos.y * pos.y
     const w2 = varInfo.weight * varInfo.weight
-    const signY = select(1.0, -1.0, r2 <= w2)
+    const signY = select(f32(1.0), f32(-1.0), r2 <= w2)
     return vec2f(pos.x, signY * pos.y)
   },
   'general',
